requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
Werkzeug
PyJWT
